requirements.txt
pymongo
python-dotenv
apache-airflow-providers-mongo